// apps/web/src/components/animations/BoxTransformation.d.ts
import React from 'react';
interface BoxTransformationProps {
    duration?: number;
    isVisible: boolean;
    onComplete?: () => void;
    seedDestination?: {
        x: number;
        y: number;
    };
}
export declare const BoxTransformation: React.FC<BoxTransformationProps>;
export default BoxTransformation;
//# sourceMappingURL=BoxTransformation.d.ts.map